@@ -1,5 +1,5 @@
-const yariPackageJson = require("@mdn/yari/package.json");
-const thisPackageJson = require("../package.json");
+import yariPackageJson from "@mdn/yari/package.json" assert { type: "json" };
+import thisPackageJson from "../package.json" assert { type: "json" };
 
 const availableYariVersion = thisPackageJson.dependencies["@mdn/yari"];
 const installedYariVersion = yariPackageJson.version;
